make less public
