+++ narwhals/_arrow/expr.py
@@ -22,9 +22,9 @@ if TYPE_CHECKING:
     from narwhals._compliant.typing import AliasNames
     from narwhals._compliant.typing import EvalNames
     from narwhals._compliant.typing import EvalSeries
+    from narwhals._compliant.typing import ScalarKwargs
     from narwhals._expression_parsing import ExprMetadata
     from narwhals.typing import RankMethod
-    from narwhals.typing import ScalarKwargs
     from narwhals.utils import Version
     from narwhals.utils import _FullContext
