+++ narwhals/_compliant/expr.py
@@ -59,6 +59,7 @@ if TYPE_CHECKING:
     from narwhals._compliant.typing import AliasNames
     from narwhals._compliant.typing import EvalNames
     from narwhals._compliant.typing import EvalSeries
+    from narwhals._compliant.typing import ScalarKwargs
     from narwhals._expression_parsing import ExprKind
     from narwhals._expression_parsing import ExprMetadata
     from narwhals.dtypes import DType
@@ -67,7 +68,6 @@ if TYPE_CHECKING:
     from narwhals.typing import NumericLiteral
     from narwhals.typing import RankMethod
     from narwhals.typing import RollingInterpolationMethod
-    from narwhals.typing import ScalarKwargs
     from narwhals.typing import TemporalLiteral
     from narwhals.typing import TimeUnit
     from narwhals.utils import Implementation
